New Grount Zero: Before Building Matrices
After a break of about 10 months from working on the project, work resumes.
I've divided the work in two levels:
1. Relevency (in time).
2. Recency.
In the first level of division we have the "Current-QA" folder, where work is continued towards a stable version. Here the most relevant files are kept, and the main assignment is building the analysis matrices.
The "Past versions" folder contains previous versions of the pipeline code. This includes "Scripts-Old", a folder that was opened when I started this repository a year ago. Another folder is "Scripts-Source" taken from a more coding step, more or less equal to the stage of the "Current-QA" folder.
If problems arise in "Current-QA", look in "Scripts-Source-Old" for possible solutions.
I am now about to start working on building the matrices.