Beginning of S03; Fixes in S02
S03:
1. Setting parameters: Attributes and values - Implemented. Not tested.
2. Querying on CHANNELLOG into channelLogQ - Implemented with logical indexing. Not tested.
3. Creating results table with proper parameters:
I decided to use outerjoin on keys to merge all the matrices from the respected participants who survived the query. This means a loop will be used, starting from the 2nd row of channelLogQ until the end. The 1st row is base of channelLogQ will be base.
ERROR: While checking matrix structure of cleaned databases - they were empty. The reason follows:
S02:
The 'cleaned' folder after channel was cleaned was not populated due to problematic a false conditoin on the preceding 'if' statement, to recognize channel content. It was fixed.
Verdict:
1. Run S02.
2. Finish and run S03.